PT-11036 -  Change data type of tax rate to float

+++ Components/Order/Hydrator/PositionHydrator.php
@@ -26,7 +26,7 @@ class PositionHydrator
         $positionStruct->setName((string) $data['articleName']);
         $positionStruct->setQuantity((int) $data['quantity']);
         $positionStruct->setStatusId((int) $data['statusId']);
-        $positionStruct->setTaxRate((int) $data['taxRate']);
+        $positionStruct->setTaxRate((float) $data['taxRate']);
         $positionStruct->setTaxId((int) $data['taxId']);
         $positionStruct->setPrice((float) $data['price']);
         $positionStruct->setTotal((float) $data['total']);
